Add team selection for spectators
CharacterManager now uses CharacterDisplayCriteria to determine whether or not to display a character on the map. When a person selects a team, the appropriate criteria is selected so that person can only see what their team is supposed to see (ie. a Ghost's team can only see Pacman if he has eaten a Power Pill).